\name{Internal functions}
\alias{addGUI}
\alias{align}
\alias{alignGUI}
\alias{alignGUI2}
\alias{anovaGUI}
\alias{askquitGUI}
\alias{baseline}
\alias{baselineGUI}
\alias{best.nmds}
\alias{best.nmdsGUI}
\alias{best.ordination}
\alias{best.ordinationGUI}
\alias{binary}
\alias{binaryGUI}
\alias{change.nameGUI}
\alias{clearconsole}
\alias{convertGUI}
\alias{correlationGUI}
\alias{delete.background}
\alias{delete.backgroundGUI}
\alias{deletefactGUI}
\alias{deleteGUI}
\alias{dendoGUI}
\alias{des.univGUI}
\alias{diversities}
\alias{diversitiesGUI}
\alias{ffcorGUI}
\alias{globalanosimGUI}
\alias{heat.map}
\alias{heat.mapGUI}
\alias{importGUI1}
\alias{importGUI2}
\alias{importGUI3}
\alias{iterative.test}
\alias{iterative.testGUI}
\alias{iterative.testGUI1}
\alias{loadGUI}
\alias{make.peak}
\alias{make.peakGUI}
\alias{newdist}
\alias{newrdata}
\alias{nmds.contour}
\alias{normalisation}
\alias{normalisationGUI}
\alias{oppdf}
\alias{pcaGUI}
\alias{peakparameters}
\alias{permanovaGUI}
\alias{plot2dGUI}
\alias{plot3dim}
\alias{plot3dimGUI}
\alias{plotord3dGUI}
\alias{plotordGUI}
\alias{pwanosim}
\alias{pwanosimGUI}
\alias{rollball}
\alias{roxdefault}
\alias{roxnewGUI}
\alias{saveasGUI}
\alias{saveGUI}
\alias{simper}
\alias{simperGUI}
\alias{simperGUI1}
\alias{trunck}
\alias{trunckGUI}
\alias{withingroupvariability}
\alias{withingroupvariabilityGUI}
\title{Internal functions for the SSCP package}
\description{
Internal functions for the SSCP package
}
\details{
These are not to be called by the user
}

\author{Rory Michelland \email{rory.michelland@gmail.com}\cr
Laurent Cauquil \email{laurent.cauquil@toulouse.inra.fr}}
\seealso{\code{\link{StatFingerprints}}}
\keyword{multivariate}